# src/middelware/tsoa/swagger.json
{
	"components": {
...
		"schemas": {
			"InternalServerError": {
				"properties": {
					"stack": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"Scope_ScopeGroup_Request": {
				"properties": {
					"scope_group_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"scope_group_id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ScopeRequest": {
				"properties": {
					"scope_id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"scope_groups": {
						"items": {
							"$ref": "#/components/schemas/Scope_ScopeGroup_Request"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			},
			"oauth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://localhost:3000/oidc/auth",
						"scopes": {
							"openid": "modify things",
							"profile": "read things",
							"email": "Get user info"
						}
					}
				}
			},
			"openid": {
				"type": "openIdConnect",
				"openIdConnectUrl": "http://localhost:3000/oidc/.well-known/openid-configuration"
			},
			"google": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://localhost:7000/auth/google",
						"scopes": {}
					}
				}
			},
			"facebook": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://localhost:7000/auth/facebook",
						"scopes": {}
					}
				}
			}
...
	},
	"info": {
		"title": "micro-identity",
		"version": "0.0.1",
		"description": "Boilerplate for creating identity microservice",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Martin Greyling"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/scope/{scope_id}": {
			"get": {
				"operationId": "GetScope",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"User API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Scopes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "scope_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "detailed",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteScope",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"User API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Scopes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "scope_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "softDelete",
						"required": false,
						"schema": {
							"default": true,
							"type": "boolean"
						}
					}
				]
			}
		},
		"/scope": {
			"get": {
				"operationId": "GetScopes",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"User API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Scopes"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "detailed",
						"required": false,
						"schema": {
							"default": false,
							"type": "boolean"
						}
					}
				]
			},
			"post": {
				"operationId": "PostScope",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"User API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Scopes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ScopeRequest"
							}
						}
					}
				}
			},
			"put": {
				"operationId": "PutScope",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"User API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Scopes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ScopeRequest"
							}
						}
					}
				}
			}
		},
		"/scope/default_scopes": {
			"post": {
				"operationId": "AddDefaultScope",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"User API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"Scopes"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/user/scopes/{user_id}": {
			"get": {
				"operationId": "GetUserScopes",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"User API Internal Server Error": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalServerError"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "user_id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}